requirements/dev.txt
-r base.txt
django-debug-toolbar==5.0.1